refactor(sdk): Restructure query builder and enhance client configuration
