+++ core-web/libs/sdk/client/src/lib/client/content/builders/query/lucene-syntax/Operand.ts
